+++ hydra-node/src/Hydra/Ledger/Cardano/Configuration.hs
@@ -7,17 +7,16 @@ module Hydra.Ledger.Cardano.Configuration (
 import Hydra.Cardano.Api
 import Hydra.Prelude
 
+import Cardano.Ledger.BaseTypes (Globals (..), boundRational, mkActiveSlotCoeff)
 import qualified Cardano.Ledger.BaseTypes as Ledger
+import Cardano.Ledger.Shelley.API (computeRandomnessStabilisationWindow, computeStabilityWindow)
 import qualified Cardano.Ledger.Shelley.Genesis as Ledger
 import qualified Cardano.Ledger.Shelley.LedgerState as Ledger
 import qualified Cardano.Ledger.Shelley.Rules.Ledger as Ledger
-import Cardano.Slotting.EpochInfo (hoistEpochInfo)
+import Cardano.Slotting.EpochInfo (fixedEpochInfo)
 import Cardano.Slotting.Time (mkSlotLength)
-import Control.Arrow (left)
-import Control.Monad.Trans.Except (runExcept)
 import qualified Data.Aeson as Json
 import qualified Data.Aeson.Types as Json
-import qualified Ouroboros.Consensus.HardFork.History as Consensus
 
 -- * Helpers
 
@@ -27,31 +26,58 @@ readJsonFileThrow parser filepath = do
   case Json.parseEither parser value of
     Left e -> fail e
     Right a -> pure a
+
 -- * Globals
 
 shelleyGenesisFromJson :: Json.Value -> Json.Parser (Ledger.ShelleyGenesis LedgerEra)
 shelleyGenesisFromJson = parseJSON
 
--- FIXME: We should not need the shelley genesis as the Ledger.Globals are
--- irrelevant (e.g. active slot coefficient) or can be derived from chain config
--- (e.g. networkId). Consequence of this would be less configuration for the
--- hydra-node (only protocol-parameters).
-newGlobals :: Ledger.ShelleyGenesis LedgerEra -> Ledger.Globals
-newGlobals shelleyGenesis =
-  Ledger.mkShelleyGlobals
-    shelleyGenesis
-    epochInfo
-    maxProtocolVersion
+data GlobalsTranslationError = GlobalsTranslationError deriving (Eq, Show)
+
+instance Exception GlobalsTranslationError
+
+-- | Create new L2 ledger 'Globals' from 'GenesisParameters'.
+newGlobals :: MonadThrow m => GenesisParameters -> m Globals
+newGlobals genesisParameters = do
+  case mkActiveSlotCoeff <$> boundRational protocolParamActiveSlotsCoefficient of
+    Nothing -> throwIO GlobalsTranslationError
+    Just slotCoeff -> do
+      let k = fromIntegral protocolParamSecurity
+      pure $
+        Globals
+          { activeSlotCoeff = slotCoeff
+          , epochInfo
+          , maxKESEvo = fromIntegral protocolParamMaxKESEvolutions
+          , maxLovelaceSupply = fromIntegral protocolParamMaxLovelaceSupply
+          , maxMajorPV
+          , networkId = toShelleyNetwork protocolParamNetworkId
+          , quorum = fromIntegral protocolParamUpdateQuorum
+          , randomnessStabilisationWindow = computeRandomnessStabilisationWindow k slotCoeff
+          , securityParameter = k
+          , slotsPerKESPeriod = fromIntegral protocolParamSlotsPerKESPeriod
+          , stabilityWindow = computeStabilityWindow k slotCoeff
+          , systemStart = SystemStart protocolParamSystemStart
+          }
  where
+  GenesisParameters
+    { protocolParamSlotsPerKESPeriod
+    , protocolParamUpdateQuorum
+    , protocolParamMaxLovelaceSupply
+    , protocolParamSecurity
+    , protocolParamActiveSlotsCoefficient
+    , protocolParamSystemStart
+    , protocolParamNetworkId
+    , protocolParamMaxKESEvolutions
+    , protocolParamEpochLength
+    , protocolParamSlotLength
+    } = genesisParameters
   -- NOTE: This is used by the ledger to discard blocks that have a version
   -- beyond a known limit. Or said differently, unused and irrelevant for Hydra.
-  maxProtocolVersion = 0
-  epochInfo =
-    Consensus.neverForksSummary epochSize slotLength
-      & Consensus.summaryToEpochInfo
-      & hoistEpochInfo (left show . runExcept)
-  epochSize = Ledger.sgEpochLength shelleyGenesis
-  slotLength = mkSlotLength (Ledger.sgSlotLength shelleyGenesis)
+  maxMajorPV = 0
+  -- NOTE: uses fixed epoch info for our L2 ledger
+  epochInfo = fixedEpochInfo protocolParamEpochLength slotLength
+  slotLength = mkSlotLength protocolParamSlotLength
+
 -- * LedgerEnv
 
 protocolParametersFromJson :: Json.Value -> Json.Parser ProtocolParameters